+++ resources/views/pages/show.blade.php
@@ -88,4 +88,17 @@
     }
 </style>
 @endpush
+
+{{-- Sidecar assets --}}
+@if(!empty($page['styles']))
+@push('styles')
+<style>{!! $page['styles'] !!}</style>
+@endpush
+@endif
+
+@if(!empty($page['scripts']))
+@push('scripts')
+<script>{!! $page['scripts'] !!}</script>
+@endpush
+@endif
 @endsection